Add video model (#23)
* Create video model

* Add video resources to router

* Video: only allow certain states in changeset

* Video: Fix order of validation

First, we want to make sure that all required attributes are given.
Afterwards, we validate for the state.

Without this fix, the function could have never been called at the end
of the validation chain, since it returned nil when the changeset was
invalid.

* Video validation: small refactor

* Introduce the /backend-scope for video management

The auto generated views are suitable for a backend, but not for the
frontend. So why not use them accordingly?

* Move VideoController to VideoBackendController

Also move view and templates (and tests).

* Use a scoped module approach over complex naming

`Backend.VideoController` is better scoped than `VideoBackendController`!

* Move files for the updated scope

* Test refactor

Use Enum.each!